+++ src/main/java/org/chainoptim/features/factory/dto/UpdateFactoryInventoryItemDTO.java
@@ -9,6 +9,6 @@ public class UpdateFactoryInventoryItemDTO {
     private Integer factoryId;
     private Integer productId;
     private Integer componentId;
-    private Double quantity;
-    private Double minimumRequiredQuantity;
+    private Float quantity;
+    private Float minimumRequiredQuantity;
 }
